Simplifying an algorithm so I can remove some code.
I'm sure the regex is the scary bit here, but let me try to mount a
defense.  First, this is about the dumbest regex you can imagine.
Here's what the code actually produces:

```ruby
ROMAN_CHUNK_REGEX = /(?:IV|CM|CD|IX|XC|XL|L|X|C|V|D|I|M)/
```

I probably should have just hardcoded it and saved readers a heart
attack.  I decided it would be nice to build it from the `Hash` though,
avoiding duplication.  That's the reason I added so many lines of code.
The trickiest bit is where I sort the chunks so two character chunks
will be tried before one character bits.

Finally, in defense of the switch to regex in general I would say that
it's a perfect fit for this problem.  A regex is a specialized tool for
dividing up some characters by meaning.  That's **exactly** what we are
trying to do here.  It's the right tool for this job.

I feel like the fact that it allowed me to turn three methods into one
defends that somewhat.  And it will do more in a second…